+++ switch/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Unreleased
 
-### 0.0.1+0
+### 0.0.1
 * Initial version
+* External switch handling
+* OpenThread connection
+* FOTA service
+* Provisioning stored in settings
+* Discovery of actuators